fix(统计报表-人员班组和检修): 修改检修的时长

+++ aiurt-module-inspection/aiurt-module-inspection-api/src/main/java/com/aiurt/boot/task/dto/OverhaulStatisticsDTO.java
@@ -62,7 +62,7 @@ public class OverhaulStatisticsDTO {
     @TableField(exist = false)
     @Excel(name = "总检修时长", width = 15)
     @ApiModelProperty(value = "总检修时长")
-    private BigDecimal maintenanceDuration;
+    private Integer maintenanceDuration;
 
     /**计划检修总数*/
     @TableField(exist = false)
